requirements.txt
websocket-client~=1.3.3
dydx-v3-python
